Continued to expand and update various docstrings for documentation.
Added eventlog for when service stops, reworked how service is started to clean up amount of functions.

Added functions to add the kab, the kabs drawers, the tools, and the tools in the kabs drawers to the backend if they do not exist, this should reduce reliance on external tools for exporting database and should reduce the occurances of duplicate tools.

cleaned up various commented out code as they are no longer needed.